+++ server/crates/arachnea-scrapyfy/src/scrapyfy/post_processes/extract_regex_items.rs
@@ -97,7 +97,10 @@ pub(super) fn apply(
     }
 
     let target_path = split_path(target);
-    root.set_output_type(&target_path, crate::scrapyfy::ScraperOutputType::ObjectArray);
+    root.set_output_type(
+        &target_path,
+        crate::scrapyfy::ScraperOutputType::ObjectArray,
+    );
     for item in extracted_items {
         root.push_node_typed(
             &target_path,